requirements.txt
flask>=2.3.0
youtube-transcript-api>=0.6.0
pytube>=15.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
langchain>=0.0.300
transformers>=4.30.0
torch>=2.0.0
google-api-python-client>=2.100.0
numpy>=1.24.0